1. loadGrid fills missing values from netcdf file 2. loadGrid can force a utm zone 3. changed max to 1e6 in precipitation

+++ smrf/framework/model_framework.py
@@ -182,8 +182,9 @@ class SMRF():
 
         # clean up the WORKDIR
         if hasattr(self, 'topo'):
-            if os.path.isfile(self.topo.stoporad_in_file):
-                os.remove(self.topo.stoporad_in_file)
+            if self.topo.stoporad_in_file is not None:
+                if os.path.isfile(self.topo.stoporad_in_file):
+                    os.remove(self.topo.stoporad_in_file)
         if hasattr(self, 'distribute'):
             if 'solar' in self.distribute.keys():
                 if os.path.isfile(self.distribute['solar'].vis_file):
